feat(information): add new page for adding information
A new page has been added to the information section of the stock-center
app. This page includes an editable editor component, allowing users to
add new information. The page is only accessible to users with private
access and the roles of 'user-user' and 'content-admin'. This feature
enhances the app's functionality by providing a dedicated space for
users to contribute new information.