feat(linkLists.ts): add new file to manage link properties and lists
The new file linkLists.ts is added to manage link properties and lists.
This file contains the definition of the LinkProperties type and several
arrays of links (downloadLinks, infoLinks, materialsLinks, miscLinks).
This change is done to centralize the management of links and improve
the maintainability of the code. Now, if a link needs to be added,
removed, or modified, it can be done in one place.