feat(home/editable.tsx): add new EditableHomepage component for content admins
A new EditableHomepage component has been added to the home directory.
This component is designed to be the main homepage component for DSC. It
includes a button for content admins to edit the content directly from
the homepage. This change is made to improve the user experience for
content admins by providing them with a more convenient way to manage
the website content.