feat(ContentView.tsx): add ContentView component for displaying content fetched from GraphQL API
The ContentView component is added to the EditablePages feature. It imports the Editor component from the "editor" package and the ContentBySlugQuery type from the "dicty-graphql-schema" package. The component takes in a data prop of type ContentBySlugQuery["contentBySlug"] and renders an Editor component with the content fetched from the GraphQL API. The Editor component is set to be non-editable and receives the storageKey and editorState from the data prop.