feat(EditablePages): add EditButton component
The EditButton component is added to the EditablePages feature. This component imports the Button component from the Material-UI library and the useNavigate hook from react-router-dom. It renders a button that, when clicked, triggers the navigate function to redirect to the "edit" route. This button will be used for editing pages in the EditablePages feature.