feat(ui-user): add Avatar component to ui-user package
A new Avatar component has been added to the ui-user package. This
component will be used to display user avatars across the application.
The Avatar component uses the MoodSharp icon from Material UI and is
wrapped in an IconButton. The component is also exported in the
package's index file for easy import elsewhere in the application.