feat(ui-user): add Information component to display user details and edit button
The Information component was added to the ui-user package. This
component displays the user's name, email, and phone number, and
includes an 'Edit' button that navigates to the user edit page. This
component enhances the user interface by providing a clear and concise
view of the user's information and an easy way to navigate to the edit
page.