feat(show.tsx): import and use displayOnAuthorzied, matchEntries from auth
The changes were made to enhance the user experience by displaying the
EditablePagesList only to users with the 'content-admin' role. This is
achieved by using the displayOnAuthorzied and matchEntries functions
from the auth module. The useStyles invocation was moved to improve code
readability and maintain a logical order of variable declarations.