feat(Title.tsx): create new Title component to encapsulate Typography settings for DCR user profile title
The Typography settings for the DCR user profile title were encapsulated
into a new Title component. This change improves code reusability and
maintainability by allowing the same title styling to be used across
multiple components or pages. The Title component is also exported from
the ui-user package, making it available for use in other parts of the
application.