feat(ui-user): add NoPages component to display when user has no editable pages
The 'Nothing to display' text for unauthorized users in user/show.tsx
has been replaced with a new NoPages component. This component provides
a more descriptive message and a better user experience. The NoPages
component has been added to the ui-user package and is exported for use
in other parts of the application.