feat(FooterWithAuth.tsx): add FooterWithAuth component to display a footer based on user authorization status
The FooterWithAuth component is added to the app's layout. It imports the Footer component from the "@dictybase/footer" package and other necessary dependencies. The component uses the useAuthorization hook to check if the user is authorized and isLoading to handle loading state. If isLoading is true, a Loader component is displayed. If isLoading is false, the displayOnAuthorzied function is used to conditionally render the Footer component based on the user's authorization status. If the user is authorized, the Footer component is rendered with data from the authFooterData.json file. If the user is unauthorized, a default Footer component is rendered.