chore(authNavbarData.json): update links in authNavbarData.json to reflect changes in route paths
The links in the authNavbarData.json file have been updated to reflect changes in the route paths. The "link" property for each item in the "items" array has been changed from "/explore/[category]/navbar" to "/explore/[category]/editable". This ensures that the links point to the correct routes in the application.