feat(header): add defaultIconItems file to define default icon items for the header component
The defaultIconItems file is added to the header package. It exports an array of objects that define the default icon items for the header component. Each object contains properties such as href, title, key, and Icon, which are used to render the icon items in the header. This file allows for easy customization and configuration of the default icon items in the header component.