refactor(edit.tsx): remove unused import and simplify props usage in components
The import statement for `useStyles` from `@dictybase/ui-dsc` is removed as it is not being used in the code. The `isEditable` prop in the `IntroEditor` component and the `isAuthorized` prop in the `CatalogLinks` component are simplified by removing the unnecessary boolean values. This improves code readability and reduces unnecessary code clutter.