feat(useSlug.ts): add useSlug hook to generate slugs based on route parameters
The useSlug hook is added to the project to generate slugs based on the route parameters. It uses the useParams hook from react-router-dom to get the name and subname parameters from the current route, and then calls the getSlug function to generate the slug. This hook can be used in components to easily generate slugs for dynamic routing.