feat(frontpage): add show.tsx files for nested routes [section]/[name]/[subname] and [section]/[name]
The show.tsx files are added to support the routes [section]/[name]/[subname]/show and [section]/[name]/show in the frontpage application. These files import the Show component from the EditablePages feature and define the access level as public using the ACCESS constant from the routes/types module. This allows the pages to be displayed publicly.