refactor(useSlug): move useSlug hook from utils folder to hooks folder
The new useSlug hook is added to the project to retrieve the slug based on the URL parameters. It uses the useParams and useLocation hooks from react-router-dom to get the necessary parameters. The hook is moved from the utils folder to the hooks folder for better organization and separation of concerns.