feat(getLastPathSegment.ts): add utility function to get the last segment of a path
The `getLastPathSegment` utility function has been added to the `common/utils` directory. This function takes a `pathname` string as input and returns the last segment of the path. It uses functional programming libraries such as `fp-ts` to split the path, filter out empty segments, exclude specific view modes, and retrieve the last segment. If no last segment is found, an empty string is returned. This utility function can be used to extract the last part of a URL path for various purposes in the application.