feat(useFetchPublications.ts): add useFetchPublications hook to fetch publication data from PubMed RSS feed
The useFetchPublications hook is added to the common/hooks directory. It fetches publication data from the PubMed RSS feed using the fetch API. If the fetch is successful, it returns the response data as a Right Either. If there is an error, it returns the error message as a Left Either.