feat(LatestPapersView.tsx): add LatestPapersView component to display a list of latest papers
The LatestPapersView component is added to display a list of latest papers. It receives an array of PublicationItem objects as data and renders a SinglePaper component for each item in the array. The SinglePaper component displays the title of each paper. This component is intended to be used in the frontpage feature of the Dicty application.